backend/requirements.txt
fastapi
uvicorn
celery
redis
sqlalchemy
alembic
pydantic-settings
httpx
